Move Plugin.description to extension; Extract PluginDescription
